refc: Renaming folder to create python package
