+++ proctai/reviewer/code_review.py
@@ -1,7 +1,7 @@
-from cloudcode.helpers import output, parser
+from proctai.helpers import output, parser
 from typing import Optional
-from cloudcode.llms.provider import LLMProvider
-from cloudcode.llms.prompts import (
+from proctai.llms.provider import LLMProvider
+from proctai.llms.prompts import (
     CODE_REVIEW_PROMPT,
     CODE_REVIEW_SYSTEM_PROMPT,
     PR_DESCRIPTION_PROMPT,
